structuring and content (12.1-12.5)
Sections are now each their own .html and the dynamic TOC now accommodates this change. Headings of sections are dynamic. Multipart examples and property lists are styled. Folder specific macros and page specific macros can be defined easily. Content up to 12.5 has been added, although most of it has not been updated. And the content looks not so great with examples back to back. It will look fine once I add solutions and graphics, but it still will not be in printable state, yet. Definitions are not styled correctly yet. There's no navigation bar.